requirements.txt
pip~=23.1.2
requests~=2.29.0
bs4~=0.0.1
pymongo~=4.3.3
python-dateutil~=2.8.2
selenium~=4.9.0
webdriver_manager~=3.8.6
pydantic~=1.10.7
beautifulsoup4~=4.12.2
python-jose[cryptography]>=3.1.0
passlib[bcrypt]>=1.7.4
httpx==0.24.1
lxml
python-crontab